1.0.1rc2 (2019-11-30) ---------------------

+++ src/pyc_wheel/_pyc_wheel.py
@@ -8,77 +8,92 @@
 import sys
 import os
 import platform
+import distutils
+import stat
 import shutil
+import tempfile
 import compileall
 import zipfile
 import hashlib
 import csv
-from base64 import urlsafe_b64encode
+import base64
+from datetime import datetime
 from pathlib import Path
 
 __all__ = ('convert_wheel',)
 
 
-HASH_TYPE = "sha256"
+HASH_ALGORITHM = hashlib.sha256
 
 
-def convert_wheel(whl_file: Path, *, with_backup=False):
-    """Generate a new whl with only pyc files.
-
-    This whl will append .compiled to the version information.
-    """
+def convert_wheel(whl_file: Path, *, with_backup=False, quiet=False):
+    """Generate a new whl with only pyc files."""
 
     if whl_file.suffix != ".whl":
         raise TypeError("File to convert must be a *.whl")
 
-    whl_path = whl_file.with_suffix("")
-    whl_dir = str(whl_path)
-
-    # Clean up leftover files
-    shutil.rmtree(whl_dir, ignore_errors=True)
-
-    # Extract our zip file temporarily
-    with zipfile.ZipFile(str(whl_file), "r") as whl_zip:
-        whl_zip.extractall(whl_dir)
-
+    whl_dir  = tempfile.mkdtemp() 
+    whl_path = Path(whl_dir)
     try:
-        # Replace all py files with pyc files
-        compileall.compile_dir(whl_dir)
+        # Extract our zip file temporarily
+        with zipfile.ZipFile(str(whl_file), "r") as whl_zip:
+            whl_zip.extractall(whl_dir)
+            members = [member for member in whl_zip.infolist()
+                       if (member.is_dir() or
+                           not member.filename.endswith(".py"))]
+
+        # Compile all py files
+        compileall.compile_dir(whl_dir,
+                               quiet=int(quiet), force=True, legacy=True)
 
         # Remove all original py files
         for py_file in whl_path.glob("**/*.py"):
-            print("Deleting py file: {}".format(py_file))
-            py_file.unlink()
+            if py_file.is_file():
+                if not quiet: print("Deleting py file: {!s}".format(py_file))
+                py_file.chmod(stat.S_IWUSR)
+                py_file.unlink()
 
         for root, dirs, files in os.walk(whl_dir):
-            for f in files:
-                if f.endswith(".py"):
-                    py_file = Path(root)/f
-                    print("Removing file: {}".format(py_file))
+            for fname in files:
+                if fname.endswith(".py"):
+                    py_file = Path(root)/fname
+                    if not quiet: print("Removing file: {!s}".format(py_file))
+                    py_file.chmod(stat.S_IWUSR)
                     py_file.unlink()
 
-        # Update the record data
-        dist_info_name = "{}.dist-info".format("-".join(whl_dir.split("-")[:-3]))
-        dist_info_path = whl_path/dist_info_name
+        for member in members:
+            file_path = whl_path/member.filename
+            timestamp = datetime(*member.date_time).timestamp()
+            try:
+                os.utime(str(file_path), (timestamp, timestamp))
+            except:
+                pass # ignore errors
+
+        dist = "-".join(whl_file.stem.split("-")[:-3])
+        dist_info_path = whl_path/"{}.dist-info".format(dist)
         rewrite_dist_info(dist_info_path)
 
         # Rezip the file with the new version info
-        if with_backup:
-            whl_file.replace(whl_file.with_suffix(whl_file.suffix + ".bak"))
-        whl_file_zip = whl_file.with_suffix(".zip")
+        whl_file_zip = whl_path.with_suffix(".zip")
         if whl_file_zip.exists(): whl_file_zip.unlink()
         shutil.make_archive(whl_dir, "zip", root_dir=whl_dir)
-        whl_file_zip.replace(whl_file)
+        if with_backup:
+            whl_file.replace(whl_file.with_suffix(whl_file.suffix + ".bak"))
+        shutil.move(str(whl_file_zip), str(whl_file))
     finally:
         # Clean up original directory
-        shutil.rmtree(whl_dir)
+        shutil.rmtree(whl_dir, ignore_errors=True)
 
 
 def rewrite_dist_info(dist_info_path: Path):
     """Rewrite the record file with pyc files instead of py files."""
 
     whl_path = dist_info_path.resolve().parent
+
+    # Rewrite the record file with pyc files instead of py files.
+
     record_path = dist_info_path/"RECORD"
+    record_path.chmod(stat.S_IWUSR)
 
     record_data = []
     with record_path.open("r") as record:
@@ -86,34 +101,65 @@ def rewrite_dist_info(dist_info_path: Path):
             if file_dest.endswith(".py"):
                 # Do not keep py files, replace with pyc files
                 file_dest = Path(file_dest)
-                pyc_fname = "{}.{}-{}{}.pyc".format(
-                                file_dest.stem,
-                                platform.python_implementation().lower(),
-                                sys.version_info.major,
-                                sys.version_info.minor)
-                pyc_file = file_dest.parent/"__pycache__"/pyc_fname
+                #pyc_fname = "{}.{}-{}{}.pyc".format(
+                #                file_dest.stem,
+                #                platform.python_implementation().lower(),
+                #                sys.version_info.major,
+                #                sys.version_info.minor)
+                #pyc_file = file_dest.parent/"__pycache__"/pyc_fname
+                pyc_file = file_dest.with_suffix(".pyc")
                 file_dest = str(pyc_file)
 
-                file_hash = hashlib.new(HASH_TYPE)
-                file_len = 0
                 pyc_path = whl_path/pyc_file
                 with pyc_path.open("rb") as f:
-                    while True:
-                        data = f.read(1024)
-                        if not data:
-                            break
-                        file_hash.update(data)
-                        file_len += len(data)
-
-                hash_str = file_hash.digest()
-                hash_str = urlsafe_b64encode(file_hash.digest())
-                hash_str = urlsafe_b64encode(file_hash.digest()).rstrip(b"=")
-                file_hash = "{}={}".format(HASH_TYPE, hash_str)
+                    data = f.read()
+                file_hash = HASH_ALGORITHM(data)
+                file_hash = "{}={}".format(file_hash.name,
+                                           _b64encode(file_hash.digest()))
+                file_len  = len(data)
             record_data.append((file_dest, file_hash, file_len))
 
     with record_path.open("w", newline="\n") as record:
         csv.writer(record,
                    lineterminator="\n").writerows(sorted(set(record_data)))
+
+    # Rewrite the wheel info file.
+
+    wheel_path = dist_info_path/"WHEEL"
+    wheel_path.chmod(stat.S_IWUSR)
+
+    with wheel_path.open("r") as wheel:
+        wheel_data = wheel.readlines()
+
+    tags = [line.split(" ")[1].strip()
+            for line in wheel_data if line.startswith("Tag: ")]
+    if not tags:
+        raise RuntimeError("No tags present in {}/{}; cannot determine target"
+                           " wheel filename".format(wheel_path.parent.name,
+                                                    wheel_path.name))
+    # Reassemble the tags for the wheel file
+    interps   = sorted({tag.split("-")[0] for tag in tags})
+    abis      = sorted({tag.split("-")[1] for tag in tags})
+    platforms = sorted({tag.split("-")[2] for tag in tags})
+
+    # [...]
+
+    with wheel_path.open("w") as wheel:
+        wheel.writelines(wheel_data)
+
+
+def _get_platform():
+    """Return our platform name 'win32', 'linux_x86_64'"""
+    result = distutils.util.get_platform().replace(".", "_").replace("-", "_")
+    if result == "linux_x86_64" and sys.maxsize == 2147483647:
+        # pip pull request #3497
+        result = "linux_i686"
+    return result
+
+
+def _b64encode(data):
+    """urlsafe_b64encode without padding"""
+    return base64.urlsafe_b64encode(data).rstrip(b"=").decode("utf-8")
 
 
 def main(args=None):
@@ -122,5 +168,10 @@ def main(args=None):
     parser.add_argument("whl_file", help="Path to whl to convert")
     parser.add_argument("--with_backup", default=False, action="store_true",
                         help="Indicates whether the backup will be created.")
+    parser.add_argument("--quiet", default=False, action="store_true",
+                        help="Indicates whether the filenames and other "
+                             "conversion information will be printed to "
+                             "standard out.")
     args = parser.parse_args(args)
-    convert_wheel(Path(args.whl_file), with_backup=args.with_backup)
+    convert_wheel(Path(args.whl_file),
+                  with_backup=args.with_backup, quiet=args.quiet)
